react-app: Extract text and currency form fields into seperate files
refs #65
