+++ react-app/src/components/common/Form/index.ts
@@ -1,0 +1,4 @@
+import { CurrencyFormField } from "./CurrencyFormField";
+import { TextFormField } from "./TextFormField";
+
+export { CurrencyFormField as Currency, TextFormField as TextInput };
